migrate to geobase

+++ lib/src/ui/map_scalebar.dart
@@ -46,7 +46,7 @@ class MapScalebar extends StatelessWidget {
     final paddedSize =
         screenSize - Offset(totalPadding.right, totalPadding.bottom) as Size;
     final scalebarCenter = alignment.alongSize(paddedSize) + paddingOffset;
-    final scalebarLat = controller.toLngLat(scalebarCenter).lat.toDouble();
+    final scalebarLat = controller.toLngLat(scalebarCenter).lat;
     final theme = Theme.of(context);
 
     final metersPerPixel = controller.getMetersPerPixelAtLatitude(
